[GMAIL GATEWAY API] removed useless interfaces for controllers
